requirements.txt
aiogram>=3.4.0
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
